langchain-course/project-2/main.py
from dotenv import load_dotenv
from typing import List
from pydantic import BaseModel, Field

load_dotenv()

from langchain.agents import create_agent
from langchain.tools import tool
from langchain.messages import HumanMessage
from langchain_openai import ChatOpenAI
from langchain_tavily import TavilySearch

class Source(BaseModel):

    """Schema for a source used by the agent"""

    url:str = Field(description="The URL of the source")

class AgentResponse(BaseModel): 
    """Schema for agent response with answer and sources"""
    answer:str = Field(description="The agent's answer to the query.")
    sources: List[Source]=Field(default_factory=list, description="List of sources used to generate the answer.")

# @tool
# def search(query: str) -> str:
#     """
#     Tool that searches over internet
#     Args:
#         query: The query to search for
#     Returns:
#         The search resullt
#     """


#     print(f"Searching for {query}")
#     return tavily.search(query=query)



llm= ChatOpenAI(model='gpt-5')
tools=[TavilySearch()]
agent= create_agent(model=llm, tools=tools, response_format=AgentResponse)



def main():
    print("Hello from project-2!")
    result=agent.invoke({"messages": HumanMessage(content="Search for 3 job positions using langchain in Budapest on Linkedin and list their details")})

    print(result)

if __name__ == "__main__":
    main()
